resolve button props circ deps
